requirements.txt

flask
python-docx
pymupdf
wordcloud
scikit-learn
matplotlib
pandas
